requirements.txt
pytz
asyncio
pyrogram
tgcrypto
python-dotenv